[Refactor] dao
